reply: render non-ASCII headers properly
When replying, purebred does not perform encoded-word decoding of
the values of the To/From/Subject header fields.  As a consequence,
raw encoded-words can appear in these values.

Use the specific header optics provided by purebred-email and remove
our buggy and unneeded helper functions.  As a result, headers with
encoded-words are presented with the values decoded.

Also add a regression test to UAT suite.